feat: align queries across engines with DISTINCT and fix quantity loading
Query alignment fixes:
- Q3: Add DISTINCT to avoid duplicate equipment from multi-relations
- Q4: Add DISTINCT for temperature points (P1=M1=O1=404 rows)
- Q8/Q9: Already use DISTINCT in CTEs, verified alignment (196 points)

Memgraph loader fix:
- Add 'quantity' field to CSV batch loading in memgraph.py
- Enables Q4/Q8/Q9 filtering on p.quantity = 'temperature'/'power'

Float32 optimization:
- Changed timeseries values from float64 to float32
- Reduces memory footprint ~18-50%

All engines now return identical results for same queries on same dataset.

+++ src/basetype_benchmark/loaders/postgres/schema_jsonb.sql
@@ -59,7 +59,7 @@ CREATE INDEX IF NOT EXISTS idx_edges_props_gin ON edges USING GIN (props jsonb_p
 CREATE TABLE IF NOT EXISTS timeseries (
     time TIMESTAMPTZ NOT NULL,
     point_id TEXT NOT NULL,
-    value DOUBLE PRECISION,
+    value REAL,
     -- Métadonnées optionnelles pour P2 (qualité, source, etc.)
     metadata JSONB DEFAULT NULL
 );
