+++ src/basetype_benchmark/loaders/postgres/timeseries.sql
@@ -7,7 +7,7 @@ CREATE TABLE IF NOT EXISTS timeseries (
     time TIMESTAMPTZ NOT NULL,
     point_id TEXT NOT NULL REFERENCES nodes(id) ON DELETE CASCADE,
     quantity TEXT NOT NULL,
-    value DOUBLE PRECISION NOT NULL,
+    value REAL NOT NULL,
     quality TEXT DEFAULT 'raw'
 );
